refactor(media): move comment-marker UI and helpers into markers/
