+++ src/lib/viewers/controls/media/markers/MarkerAvatar.tsx
