+++ src/lib/viewers/controls/media/markers/MarkerAvatarStack.scss
