+++ src/lib/viewers/controls/media/markers/MarkerAvatarStack.tsx
