+++ src/lib/viewers/controls/media/markers/MarkerCluster.scss
