+++ src/lib/viewers/controls/media/markers/__tests__/MarkerAvatar-test.tsx
