+++ src/lib/viewers/controls/media/markers/__tests__/MarkerAvatarStack-test.tsx
