+++ src/lib/viewers/controls/media/markers/helpers/__tests__/buildClusters-test.ts
@@ -1,5 +1,5 @@
 import buildClusters from '../buildClusters';
-import { CommentMarker } from '../types';
+import { CommentMarker } from '../../types';
 
 describe('buildClusters', () => {
     test('should return empty array when markers is empty', () => {
