+++ src/lib/viewers/controls/media/markers/helpers/useDismissableMarkerSelection.ts
